Textures skapas med material info, inte helt fungerande just nu vilket är anledningen att jag skapade en ny branch
