+++ DV1542-Projekt/DV1542-Projekt.vcxproj.filters
@@ -39,6 +39,9 @@
     <ClCompile Include="Shadowmap.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="OBJLoader.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <FxCompile Include="PixelShader.hlsl" />
